+++ Packages/com.unity.inputsystem/InputSystem/Devices/Remote/InputRemoting.cs
@@ -517,14 +517,12 @@ namespace UnityEngine.Experimental.Input
                 try
                 {
                     device = receiver.m_LocalManager.AddDevice(data.layout,
-                        string.Format("Remote{0}::{1}", msg.participantId, data.name));
+                        $"Remote{msg.participantId}::{data.name}");
                 }
                 catch (Exception exception)
                 {
-                    Debug.Log(
-                        string.Format(
-                            "Could not create remote device '{0}' with layout '{1}' locally (exception: {2})",
-                            data.description, data.layout, exception));
+                    Debug.LogError(
+                        $"Could not create remote device '{data.description}' with layout '{data.layout}' locally (exception: {exception})");
                     return;
                 }
                 device.m_Description = data.description;